Move shared post_params method into concern
